fix: deploy latest task definition
Update the workflow to trigger a new deploy of the
latest ACTIVE task definition using the AWS CLI.

This is being done so that the ECS service's task definition
attribute remains in sync with the Terraform.

The behaviour of the ECS service has been updated to always
deploy the latest ACTIVE task definition, thereby giving precedence
to the deployment workflows in this repo.